Doggies/v0/pyproject.toml
[tool.poetry]
name = "doggies"
version = "0.1.0"
description = "AI-powered dog adoption Telegram bot for Koh Phangan, Thailand"
authors = []
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
anthropic = "^0.40.0"
supabase = "^2.10.0"
python-telegram-bot = {extras = ["job-queue"], version = "^21.0"}
pydantic = "^2.10.0"
pydantic-settings = "^2.7.0"
pillow = "^11.0.0"
python-dotenv = "^1.0.0"
httpx = "^0.28.0"
google-api-python-client = "^2.150.0"
google-auth = "^2.36.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]
